refactor: 화면 너비 감지 로직을 useClientWidth 커스텀 훅으로 분리
